Re-emphasize JupyterHub
Our current primary product is *customized JupyterHubs*,
so let's lead with that. The secondary section has a little
bit more explanation & context.

Trimmed so that the title is < 2 lines, and the description
is also < 3 lines.

Emphasis is on the customized JupyterHub part, since that's
what we really want to communicate to our users.