Call them 2i2c managed JupyterHubs rather than hubs (#42)
* Call them 2i2c managed JupyterHubs rather than hubs

https://github.com/2i2c-org/pilot-hubs/issues/314 talks
about the need to be very explicit in what we call the
service we provide. The word 'hub' is very overloaded.

2i2c managed infrastructure has an important property that our naming
should reflect.  It isn't a *new thing* we are developing, but a
deployment of an upstream, purely open source product. This means
our users can always:

1. Take their infrastructure from 2i2c, and run it themselves
   with help from the community.
2. Work with us to *train* their own staff to run their JupyterHub,
   so they can participate fully in running the infrastructure. This
   gives them organizational control over what is built, and they
   can walk away easier if needed.
3. Any improvements we make in terms of features built or
   operations smoothed will be made in upstream repositories, and

I think directly using the word JupyterHub is a good way to
communicate this.

Alternative suggestions:

1. just 'JupyterHub'
2. 2i2c JupyterHub
3. JupyterHub managed by 2i2c

I prefer (2) if needed.

Question is, how does this related to the Jupyter Trademark
policy (https://jupyter.org/governance/trademarks.html)?
Particularly:

> Any commercial use of the Jupyter trademarks in product or company
> names must be approved first by Project Jupyter. Some uses, like calling
> a company “The Jupyter Company,” or a product “Jupyter Hosting” or
> “Jupyter Cloud” will be refused. This is because they are overly broad,
> or confusing as to whether Jupyter is open source or commercial, or
> whether your product or organization is affiliated with or sponsored by
> Project Jupyter.

(from
https://jupyter.org/governance/trademarks.html#uses-that-always-require-approval)

So no matter what, I think this needs to go to the Jupyter steering
council. I personally believe that '2i2c managed JupyterHub' is fine,
but that's not for me to decide.

* Reword the 'features' page

- Remove 'distributing content', add 'authentication' instead
- Trim content size

* Re-title infra intro section

* Shorten menu item name

* Consistency fix

* Minor grammatical fix

Co-authored-by: Chris Holdgraf <choldgraf@berkeley.edu>